制約
what
他者ユーザーの投稿記事を編集できないようにした
why
勝手に編集されたら困るから

+++ resources/views/show.blade.php
@@ -7,11 +7,15 @@
     </div>
     <div>
         <a href={{ route('blog.list') }}>一覧に戻る</a>
-        | <a href={{ route('blog.edit', ['id' =>  $blog->id]) }}>編集</a>
-        <p></p>
-        {{ Form::open(['method' => 'delete', 'route' => ['blog.destroy', $blog->id]]) }}
-            {{ Form::submit('削除') }}
-        {{ Form::close() }}
+        @auth
+          @if ($blog->user_id === $login_user_id)
+            | <a href={{ route('blog.edit', ['id' =>  $blog->id]) }}>編集</a>
+            <p></p>
+            {{ Form::open(['method' => 'delete', 'route' => ['blog.destroy', $blog->id]]) }}
+                {{ Form::submit('削除') }}
+            {{ Form::close() }}
+          @endif
+        @endauth
     </div>
   
 @endsection
